[Modify] PlayingContents 분리
